Fixed problems with asynPrint formats
git-svn-id: https://subversion.xor.aps.anl.gov/synApps/areaDetector/trunk@15799 dc6c5ff5-0b8b-c028-a01f-ffb33f00fc8b

+++ ADApp/mar345Src/mar345.cpp
@@ -266,8 +266,8 @@ asynStatus mar345::readServer(char *input, size_t maxChars, double timeout)
     lock();
     if (nread == 0) return(status);
     if (status) asynPrint(pasynUser, ASYN_TRACE_ERROR,
-                    "%s:%s, timeout=%f, status=%d received %d bytes\n%s\n",
-                    driverName, functionName, timeout, status, nread, input);
+                    "%s:%s, timeout=%f, status=%d received %lu bytes\n%s\n",
+                    driverName, functionName, timeout, status, (unsigned long)nread, input);
     /* Set output string so it can get back to EPICS */
     setStringParam(ADStringFromServer, input);
     callParamCallbacks();
